pyproject.toml
[tool.poetry]
name = "drfjson"
version = "0.1.0"
description = ""
authors = ["Alexey Goloburdin <sterx@rl6.ru>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.9"
djangorestframework = "^3.12.4"
psycopg2-binary = "^2.9.1"
django-stubs = "^1.9.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
